requirements_test.txt
-r requirements.txt

django-debug-toolbar==1.9.1
django-nose==1.4.5
django-webtest==1.9.3
webtest==2.0.30
